@@ -1,107 +1,149 @@
 {
-  "name": "wonderwp\/site",
-  "type": "project",
-  "authors": [
-    {
-      "name": "Jeremy Desvaux",
-      "email": "jeremy.desvaux@wonderful.fr",
-      "homepage": "https:\/\/github.com\/jeremy-wdf"
-    }
-  ],
-  "config": {
-    "preferred-install": {
-      "wonderwp\/*": "source",
-      "agencewonderful\/*": "source",
-      "*": "dist"
+    "name": "agencewonderful\/wonderwp-base",
+    "type": "project",
+    "description": "",
+    "authors": [
+        {
+            "name": "Jeremy Desvaux",
+            "email": "jeremy.desvaux@wonderful.fr",
+            "homepage": "https:\/\/github.com\/jeremy-wdf"
+        }
+    ],
+    "config": {
+        "preferred-install": {
+            "wonderwp\/*": "source",
+            "agencewonderful\/*": "source",
+            "*": "dist"
+        },
+        "sort-packages": true,
+        "optimize-autoloader": true,
+        "allow-plugins": {
+            "composer\/installers": true,
+            "johnpbloch\/wordpress-core-installer": true
+        }
     },
-    "sort-packages": true
-  },
-  "repositories": [
-    {
-      "type": "composer",
-      "url": "https:\/\/wpackagist.org"
+    "repositories": [
+        {
+            "type": "composer",
+            "url": "https:\/\/wpackagist.org"
+        },
+        {
+            "type": "composer",
+            "url": "https:\/\/won.wonderful.fr\/satis\/packages-mirror"
+        },
+        {
+            "type": "composer",
+            "url": "https:\/\/composer.deliciousbrains.com"
+        }
+    ],
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "require": {
+        "php": ">=8.0",
+        "ext-json": "*",
+        "agencewonderful/wonderwp-theme-core": "dev-develop@dev",
+        "agencewonderful/wwp-actu": "dev-develop@dev",
+        "agencewonderful/wwp-alerte": "dev-develop@dev",
+        "agencewonderful/wwp-branchepage": "dev-develop@dev",
+        "agencewonderful/wwp-cache": "dev-develop",
+        "agencewonderful/wwp-contact": "dev-develop@dev",
+        "agencewonderful/wwp-cron": "dev-develop@dev",
+        "agencewonderful/wwp-download": "dev-develop@dev",
+        "agencewonderful/wwp-emploi": "dev-develop@dev",
+        "agencewonderful/wwp-espace-restreint": "dev-develop@dev",
+        "agencewonderful/wwp-event": "dev-develop@dev",
+        "agencewonderful/wwp-faq": "dev-develop@dev",
+        "agencewonderful/wwp-galerie": "dev-develop@dev",
+        "agencewonderful/wwp-generator": "dev-develop@dev",
+        "agencewonderful/wwp-gutenberg-utils": "dev-develop@dev",
+        "agencewonderful/wwp-jeux": "dev-develop@dev",
+        "agencewonderful/wwp-maintenance": "dev-develop@dev",
+        "agencewonderful/wwp-map": "dev-develop@dev",
+        "agencewonderful/wwp-newsletter": "dev-develop@dev",
+        "agencewonderful/wwp-newsroom": "dev-develop@dev",
+        "agencewonderful/wwp-parcours": "dev-develop@dev",
+        "agencewonderful/wwp-partenaires": "dev-develop@dev",
+        "agencewonderful/wwp-rgpd": "dev-develop@dev",
+        "agencewonderful/wwp-search": "dev-develop@dev",
+        "agencewonderful/wwp-slider": "dev-develop@dev",
+        "agencewonderful/wwp-social": "dev-develop@dev",
+        "agencewonderful/wwp-stats": "dev-develop@dev",
+        "agencewonderful/wwp-tracker": "dev-develop@dev",
+        "agencewonderful/wwp-translator": "dev-develop@dev",
+        "agencewonderful/wwp-trombinoscope": "dev-develop@dev",
+        "agencewonderful/wwp-video": "dev-develop@dev",
+        "agencewonderful/wwp-vote": "dev-develop@dev",
+        "composer/installers": "^1.9",
+        "deliciousbrains-plugin/wp-migrate-db-pro": "^2.0",
+        "johnpbloch/wordpress": "^6.0",
+        "oscarotero/env": "^1.0",
+        "roots/wp-password-bcrypt": "1.0.0",
+        "twig/twig": "^2.0",
+        "vlucas/phpdotenv": "^2.0.1",
+        "wp-cli/wp-cli": "^2.4",
+        "wp-cli/wp-cli-bundle": "^2.4",
+        "wp-media/wp-rocket": "^3.9",
+        "wpackagist-plugin/hyperlink-group-block": "^1.0",
+        "wpackagist-plugin/imagify": "^2.1",
+        "wpackagist-plugin/secupress": "^2.2",
+        "wpackagist-plugin/simple-custom-css": "^4.0",
+        "wpackagist-plugin/wp-media-library-categories": "^2.0",
+        "wpackagist-plugin/wp-nested-pages": "^3.0"
     },
-    {
-      "type": "composer",
-      "url": "https:\/\/won.wonderful.fr\/satis\/packages-mirror"
-    }
-  ],
-  "minimum-stability": "dev",
-  "prefer-stable": true,
-  "require": {
-    "php": ">=7.2",
-    "agencewonderful/wonderwp-theme-core": "dev-develop@dev",
-    "agencewonderful/wwp-actu": "dev-develop@dev",
-    "agencewonderful/wwp-branchepage": "dev-develop@dev",
-    "agencewonderful/wwp-cache": "dev-develop",
-    "agencewonderful/wwp-contact": "dev-develop@dev",
-    "agencewonderful/wwp-download": "dev-develop@dev",
-    "agencewonderful/wwp-emploi": "dev-develop@dev",
-    "agencewonderful/wwp-espace-restreint": "dev-develop@dev",
-    "agencewonderful/wwp-faq": "dev-develop@dev",
-    "agencewonderful/wwp-generator": "dev-feature/wwp_v2@dev",
-    "agencewonderful/wwp-gutenberg-utils": "dev-develop@dev",
-    "agencewonderful/wwp-jeux": "dev-develop@dev",
-    "agencewonderful/wwp-lazyload": "dev-develop@dev",
-    "agencewonderful/wwp-map": "dev-develop@dev",
-    "agencewonderful/wwp-meteo": "dev-develop@dev",
-    "agencewonderful/wwp-newsletter": "dev-develop@dev",
-    "agencewonderful/wwp-partenaires": "dev-develop@dev",
-    "agencewonderful/wwp-rgpd": "dev-develop@dev",
-    "agencewonderful/wwp-search": "dev-develop@dev",
-    "agencewonderful/wwp-social": "dev-develop@dev",
-    "agencewonderful/wwp-stats": "dev-develop@dev",
-    "agencewonderful/wwp-tracker": "dev-develop@dev",
-    "agencewonderful/wwp-translator": "dev-develop@dev",
-    "agencewonderful/wwp-trombinoscope": "dev-develop@dev",
-    "agencewonderful/wwp-video": "dev-develop@dev",
-    "agencewonderful/wwp-vote": "dev-develop@dev",
-    "composer/installers": "~1.0.12",
-    "johnpbloch/wordpress": "^5.2",
-    "oscarotero/env": "^1.0",
-    "roots/wp-password-bcrypt": "1.0.0",
-    "twig/twig": "^2.0",
-    "vlucas/phpdotenv": "^2.0.1",
-    "wpackagist-plugin/disable-gutenberg": "^1.6",
-    "wpackagist-plugin/google-sitemap-generator": "dev-trunk",
-    "wpackagist-plugin/wp-nested-pages": "^3.0"
-  },
-  "require-dev": {
-    "phpunit\/phpunit": "^6",
-    "squizlabs\/php_codesniffer": "^2.5.1"
-  },
-  "autoload": {
-    "psr-4": {
-      "WonderWp\\Theme\\Child\\": "web\/app\/themes\/wwp_child_theme\/includes",
-      "WonderWp\\Plugin\\Jeux\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-jeux\/includes",
-      "WonderWp\\Plugin\\Contact\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-contact\/includes",
-      "WonderWp\\Plugin\\EspaceRestreint\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-espace-restreint\/includes",
-      "WonderWp\\Plugin\\RGPD\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-rgpd\/includes",
-      "WonderWp\\Plugin\\Newsletter\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-newsletter\/includes",
-      "WonderWp\\Plugin\\Video\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-video\/includes",
-      "WonderWp\\Plugin\\Trombinoscope\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-trombi\/includes",
-      "WonderWp\\Plugin\\Partenaires\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-partenaires\/includes",
-      "WonderWp\\Plugin\\Map\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-map\/includes",
-      "WonderWp\\Plugin\\GutenbergUtils\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-gutenberg-utils\/includes"
-    }
-  },
-  "extra": {
-    "installer-paths": {
-      "web\/app\/mu-plugins\/{$name}\/": [
-        "type:wordpress-muplugin"
-      ],
-      "web\/app\/plugins\/{$name}\/": [
-        "type:wordpress-plugin"
-      ],
-      "web\/app\/themes\/{$name}\/": [
-        "type:wordpress-theme"
-      ]
+    "require-dev": {
+        "phpunit\/phpunit": "^9",
+        "squizlabs\/php_codesniffer": "^2.5.1"
     },
-    "wordpress-install-dir": "web\/wp"
-  },
-  "scripts": {
-    "test": [
-      "vendor\/bin\/phpcs --ignore=web\/wp\/,vendor\/ -n -s ."
-    ]
-  }
+    "autoload": {
+        "psr-4": {
+            "WonderWp\\Theme\\Child\\": "web\/app\/themes\/wwp_child_theme\/includes",
+            "WonderWp\\Plugin\\Bp\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-branchepage\/includes",
+            "WonderWp\\Plugin\\Contact\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-contact\/includes",
+            "WonderWp\\Plugin\\Emploi\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-emploi\/includes",
+            "WonderWp\\Plugin\\EspaceRestreint\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-espace-restreint\/includes",
+            "WonderWp\\Plugin\\Event\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-event\/includes",
+            "WonderWp\\Plugin\\Faq\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-faq\/includes",
+            "WonderWp\\Plugin\\Galerie\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-galerie\/includes",
+            "WonderWp\\Plugin\\GutenbergUtils\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-gutenberg-utils\/includes",
+            "WonderWp\\Plugin\\Jeux\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-jeux\/includes",
+            "WonderWp\\Plugin\\Map\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-map\/includes",
+            "WonderWp\\Plugin\\NewsRoom\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-newsroom\/includes",
+            "WonderWp\\Plugin\\Newsletter\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-newsletter\/includes",
+            "WonderWp\\Plugin\\Parcours\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-parcours\/includes",
+            "WonderWp\\Plugin\\Partenaires\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-partenaires\/includes",
+            "WonderWp\\Plugin\\RGPD\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-rgpd\/includes",
+            "WonderWp\\Plugin\\Social\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-social\/includes",
+            "WonderWp\\Plugin\\Trombinoscope\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-trombinoscope\/includes",
+            "WonderWp\\Plugin\\Video\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-video\/includes",
+            "WonderWp\\Plugin\\Vote\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-vote\/includes",
+            "WonderWp\\Plugin\\Cron\\": "web\/app\/plugins\/wwp-cron\/includes",
+            "WonderWp\\Plugin\\Actu\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-actu\/includes",
+            "WonderWp\\Plugin\\Slider\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-slider\/includes",
+            "WonderWp\\Plugin\\Search\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-search\/includes",
+            "WonderWp\\Plugin\\Stats\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-stats\/includes",
+            "WonderWp\\Plugin\\Tracker\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-tracker\/includes",
+            "WonderWp\\Plugin\\Translator\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-translator\/includes",
+            "WonderWp\\Plugin\\Download\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-download\/includes",
+            "WonderWp\\Plugin\\Alerte\\Child\\": "web\/app\/themes\/wwp_child_theme\/plugins\/wwp-alerte\/includes"
+        }
+    },
+    "extra": {
+        "installer-paths": {
+            "web\/app\/mu-plugins\/{$name}\/": [
+                "type:wordpress-muplugin"
+            ],
+            "web\/app\/plugins\/{$name}\/": [
+                "type:wordpress-plugin"
+            ],
+            "web\/app\/themes\/{$name}\/": [
+                "type:wordpress-theme"
+            ]
+        },
+        "wordpress-install-dir": "web\/wp"
+    },
+    "scripts": {
+        "test": [
+            "vendor\/bin\/phpcs --ignore=web\/wp\/,vendor\/ -n -s ."
+        ]
+    }
 }
