@@ -3,13 +3,15 @@ describe('Smoke Test', () => {
     let host  = Cypress.env('host') || Cypress.config('host'),
         pages = Cypress.config('smoke-pages');
 
-    pages.forEach(url => {
+    if(pages) {
+      pages.forEach(url => {
         it('Checks that page ' + host + url + ' has no fatal', () => {
 
-            cy.visit(host + url);
-            cy.get("#colophon").should('be.visible');
+          cy.visit(host + url);
+          cy.checkNoFatal();
 
         });
-    });
+      });
+    }
 
 });
