@@ -7,33 +7,46 @@
  */
 ?>
 
-<div id="post-<?php the_ID(); ?>" <?php post_class('page-' . $post->post_name); ?> data-name="<?php echo $post->post_name; ?>">
+<div id="post-<?php the_ID(); ?>" <?php post_class('page-' . $post->post_name); ?>
+     data-name="<?php echo $post->post_name; ?>">
     <?php
     $postThumb = \WonderWp\Component\Media\Medias::getFeaturedImage(get_the_ID());
     ?>
-    <?php if (!empty($postThumb)) : ?>
-        <div class="post-thumbnail">
-            <?php the_post_thumbnail('large',['loading'=>'eager','class'=>'no-lazy']); ?>
-        </div><!-- .post-thumbnail -->
-    <?php endif; ?>
 
     <header class="entry-header <?php if (!empty($postThumb)) {
         echo 'hasPostThumb';
     } ?>">
-        <?php the_title('<h1 class="entry-title">', '</h1>'); ?>
-
+        <?php if (!empty($postThumb)) : ?>
+            <figure class="post-thumbnail">
+                <?php the_post_thumbnail('large', ['loading' => 'eager', 'class' => 'no-lazy']); ?>
+            </figure>
+        <?php endif; ?>
+        <?php
+        if (!\WonderWp\Functions\isAjax()) {
+            /** @var \WonderWp\Theme\Core\Service\ThemeViewService $viewService */
+            $viewService = wwp_get_theme_service('view');
+            if (!empty($viewService) && is_object($viewService)) {
+                echo $viewService->getBreadcrumbs();
+            }
+        } ?>
+        <div class="entry-title-wrapper">
+            <div class="container">
+                <?php echo '<h1 class="entry-title">' . str_replace(' | ', '<br>', get_the_title()) . '</h1>'; ?>
+                <?php
+                $excerpt = !empty($post->post_excerpt) ? $post->post_excerpt : null;
+                if (!empty($excerpt)) {
+                    //Allow shortcodes in excerpt
+                    echo '<div class="excerpt">' . apply_filters('the_excerpt', $excerpt) . '</div>';
+                }
+                ?>
+            </div>
+        </div>
     </header><!-- .entry-header -->
 
     <div class="entry-content">
         <main class="main" role="main">
 
-        <?php
-        /** @var \WonderWp\Theme\Core\Service\ThemeViewService $viewService */
-        if (!empty($viewService) && is_object($viewService)) {
-            echo $viewService->getBreadcrumbs();
-        } ?>
-
-        <?php the_content(); ?>
+            <?php the_content(); ?>
 
         </main><!-- main -->
     </div><!-- .entry-content -->
