@@ -12,36 +12,77 @@
 
 </div><!-- .site-content -->
 
-<div class="container-s">
-    <?php do_action('wwp_before_footer'); ?>
-</div>
+<?php if (!\WonderWp\Functions\isAjax()): ?>
 
-<footer id="colophon" class="site-footer" role="contentinfo">
-    <div class="footer-content">
-        <?php
-        //Menu du footer
-        //Pour fonctionner automatiquement, creer un footer appelé Footer fr_FR pour la france dans apparence/menus dans le BO
-        $footerMenuName = 'Footer ' . get_locale();
-        if (is_nav_menu($footerMenuName)) {
-            echo '<nav role="navigation" aria-label="Menu secondaire">';
-                wp_nav_menu(['menu' => $footerMenuName]);
-            echo '</nav>';
-        }
-        ?>
-        <?php
-        //Liens vers les reseaux sociaux
-        //Decommenter les 2 lignes suivantes si vous souhaitez afficher le composant des liens vers les profils sociaux adminstres depuis le BO
-        //$rs = new \WonderWp\Theme\Child\Components\ReseauxSociaux\ReseauxSociauxComponent();
-        //echo $rs->getMarkup();
-        ?>
-    </div><!-- .site-info -->
-</footer><!-- .site-footer -->
+    <div class="before-footer">
+        <div class="container-s">
+            <?php do_action('wwp_before_footer'); ?>
+        </div>
+    </div>
 
-</div><!-- .site -->
 
-<?php do_action('wwp_after_footer'); ?>
+    <footer id="colophon" class="site-footer" role="contentinfo">
 
-<?php wp_footer(); ?>
+        <div class="footer-content-top">
+            <div class="container-l">
 
-</body>
-</html>
+                <div class="first-col">
+                    <?php
+                    echo '<a href="/" class="logo" aria-label="' . trad('back.to.home', WWP_THEME_TEXTDOMAIN) . '"><img src="' . apply_filters('wwp-header-logo', '/app/themes/wwp_child_theme/assets/raw/images/logo-site-white.svg') . '" alt="Mon site - accueil"></a>'
+                    ?>
+                </div>
+
+                <?php
+                //Menu du footer
+                //Pour fonctionner automatiquement, creer un footer appelé Footer fr_FR pour la france dans apparence/menus dans le BO
+                $footerMenuName = 'Footer ' . get_locale();
+                if (is_nav_menu($footerMenuName)) {
+                    echo '<nav role="navigation" aria-label="Menu footer principal">';
+                    wp_nav_menu(['menu' => $footerMenuName]);
+                    echo '</nav>';
+                }
+                ?>
+
+                <div class="third-col">
+                    <?php
+                    //Reseaux sociaux
+                    $rs = new \WonderWp\Theme\Child\Components\ReseauxSociaux\ReseauxSociauxComponent();
+                    echo $rs->getMarkup();
+                    ?>
+                </div>
+
+            </div>
+        </div>
+
+        <div class="footer-content-bottom">
+            <div class="container-l">
+
+                <div class="first-col">
+                    <address><?php echo trad('footer.address.trad', WWP_THEME_TEXTDOMAIN) ?></address>
+                </div>
+
+                <?php
+                //Menu du footer
+                $footerMenuName = 'Footer ' . get_locale() . ' 2';
+                if (is_nav_menu($footerMenuName)) {
+                    echo '<nav role="navigation" aria-label="Menu footer secondaire">';
+                    wp_nav_menu(['menu' => $footerMenuName]);
+                    echo '</nav>';
+                }
+                ?>
+
+                <div class="third-col"></div>
+            </div>
+        </div>
+
+    </footer><!-- .site-footer -->
+
+    </div><!-- .site -->
+
+    <?php do_action('wwp_after_footer'); ?>
+
+    <?php wp_footer(); ?>
+
+    </body>
+    </html>
+<?php endif; /* isAjax */ ?>
