@@ -3,6 +3,8 @@
 namespace WonderWp\Theme\Child\Components\ReseauxSociaux;
 
 use WonderWp\Theme\Core\Component\AbstractComponent;
+use function WonderWp\Functions\array_merge_recursive_distinct;
+use function WonderWp\Functions\paramsToHtml;
 
 class ReseauxSociauxComponent extends AbstractComponent
 {
@@ -11,27 +13,36 @@ class ReseauxSociauxComponent extends AbstractComponent
 
         $markup = '';
 
-        $reseauxOpts = [
-            'facebook',
-            'twitter',
-            'youtube',
-            'instagram',
-            'pinterest',
-            'flickr',
-            'linkedin'
-        ];
+        $reseauxOpts = !empty($opts['reseaux'])
+            ? $opts['reseaux']
+            : [
+                'facebook',
+                'twitter',
+                'youtube',
+                'vimeo',
+                'instagram',
+                'pinterest',
+                'flickr',
+                'linkedin',
+                'tiktok',
+                'snapchat'
+            ];
 
         $reseauxActifs = [];
 
         foreach ($reseauxOpts as $reseau) {
             $link = get_option('wonderwp_rs_' . $reseau);
             if (!empty($link)) {
-                $reseauxActifs[] = '<li class="' . $reseau . '"><a href="' . $link . '" target="_blank">' . __('rs.follow.' . $reseau, WWP_THEME_TEXTDOMAIN) . '</a></li>';
+                $reseauxActifs[] = '<li class="' . $reseau . '"><a href="' . $link . '" target="_blank" aria-label="' . __('rs.follow.' . $reseau, WWP_THEME_TEXTDOMAIN) . ' (nouvelle fenêtre)">' . getSvgIcon($reseau) . '<span>' . __('rs.follow.' . $reseau, WWP_THEME_TEXTDOMAIN) . '<span></a></li>';
             }
         }
 
         if (!empty($reseauxActifs)) {
-            $markup = '<ul class="social-networks">
+            $params       = [
+                'class' => ['social-networks'],
+            ];
+            $passedParams = !empty($opts['listParams']) ? $opts['listParams'] : [];
+            $markup       = '<ul ' . paramsToHtml(array_merge_recursive_distinct($params, $passedParams)) . '>
                 ' . implode("\n", $reseauxActifs) . '
             </ul>';
         }
