@@ -16,10 +16,12 @@ class TabsComponent extends TwigComponent
 {
     protected $class;
     protected $content;
-    private $tabs;
+    private   $tabs;
 
-    private function setTabItems($tabItems) {
+    private function setTabItems($tabItems)
+    {
         $this->tabItems = $tabItems;
+
         return $this;
     }
 
@@ -28,17 +30,21 @@ class TabsComponent extends TwigComponent
         parent::__construct(__DIR__, 'tabs');
     }
 
-    public function addTab($title, $content, $shortcode = "") {
-        $tab = new TabItem();
-        $tab->title = $title;
+    public function addTab($title, $content, $id = "", $class = "", $shortcode = "")
+    {
+        $tab          = new TabItem();
+        $tab->title   = $title;
         $tab->content = $this->extractContent($content, $shortcode);
+        $tab->id      = $id;
+        $tab->class   = $class;
         $this->tabs[] = $tab;
     }
 
-    private function extractContent($content, $shortcode = "") {
-        if(!empty($shortcode)) {
+    private function extractContent($content, $shortcode = "")
+    {
+        if (!empty($shortcode)) {
             $shortcodes = ChildThemeShortcodeService::extractShortcode($content, $shortcode);
-            $markup = do_shortcode($shortcodes[0]);
+            $markup     = do_shortcode($shortcodes[0]);
         } else {
             $markup = $content;
         }
@@ -49,6 +55,7 @@ class TabsComponent extends TwigComponent
     public function getMarkup(array $opts = [])
     {
         $this->setTabItems($this->tabs);
+
         return parent::getMarkup($opts);
     }
 }
