@@ -26,30 +26,39 @@ class VideoComponent extends AbstractComponent
 
     /**
      * @var string
-     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Videp mp4"})
+     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Video mp4"})
      */
     protected $videoMp4;
 
     /**
      * @var string
-     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Videp ogg"})
+     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Video ogg"})
      */
-    /** @var string */
     protected $videoOgg;
 
     /**
      * @var string
-     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Videp webm"})
+     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Video webm"})
      */
-    /** @var string */
     protected $videoWebm;
 
     /**
      * @var string
-     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Image"})
+     * @BlockAttributes(component="MediaUpload",type="string",componentAttributes={"placeholder":"Image"})
      */
-    /** @var string */
     protected $image;
+
+    /**
+     * @var string
+     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Lazy"})
+     */
+    protected $lazy;
+
+    /**
+     * @var string
+     * @BlockAttributes(component="PlainText",type="string",componentAttributes={"placeholder":"Autoplay"})
+     */
+    protected $autoplay;
 
     /**
      * @param string $title
@@ -111,65 +120,102 @@ class VideoComponent extends AbstractComponent
         return $this;
     }
 
-    public function getMarkup(array $opts = [])
+    /**
+     * @param string $lazy
+     *
+     * @return VideoComponent
+     */
+    public function setLazy(string $lazy): VideoComponent
     {
+        $this->lazy = $lazy;
 
-        $index = !empty($opts['index']) ? $opts['index'] : 0;
-        $playerWrapId = 'video-player-wrap-' . $index;
+        return $this;
+    }
 
-        $markup = '
-        <div class="video-wrapper" data-video-native-component>';
+    /**
+     * @param string $autoplay
+     *
+     * @return VideoComponent
+     */
+    public function setAutoplay(string $autoplay): VideoComponent
+    {
+        $this->autoplay = $autoplay;
 
-        if (!empty($this->image)) {
-            $markup .= '<div class="video-image-wrapper">
-                <img src="' . $this->image . '" alt="">
-            </div>';
-        }
+        return $this;
+    }
 
-        $markup .= '<button class="video-trigger"><span>' . trad('video.play.label', WWP_THEME_TEXTDOMAIN) . ' ' . $this->title . '</span></button>';
+    public static $template = <<<EOD
 
-        $markup .= '
-        <div class="video-player-wrap" id="' . $playerWrapId . '">
-            <div class="video-player" id="' . $playerWrapId . '">
-
-               <video class="video" ' . (!empty($this->image) ? 'poster="' . $this->image . '"' : '') . '>';
-        if (!empty($this->videoMp4)) {
-            $markup .= '<source src="' . $this->videoMp4 . '" type="video/mp4">';
-        }
-        if (!empty($this->videoOgg)) {
-            $markup .= '<source src="' . $this->videoOgg . '" type="video/ogg">';
-        }
-        if (!empty($this->videoWebm)) {
-            $markup .= '<source src="' . $this->videoWebm . '" type="video/webm">';
-        }
-        $markup .= '
-                    ' . trad('no_video_support.trad', WWP_THEME_TEXTDOMAIN) . '
-                </video>
-
-                <div class="video-controls controls">
-                   <div class="progress-wrap">
-                      <progress class="progress" value="0" max="100">
-                         <span class="progress-bar"></span>
-                      </progress>
-                   </div>
-                   <div class="buttons">
+    {image}
+    <button class="video-trigger"><span>{label}</span></button>
+    <div class="video-player-wrap" id="{id}">
+        <div class="video-player">
+            <video class="video" {poster} {autoplay} {muted} {playsinline}>
+                {videoMp4}
+                {videoOgg}
+                {videoWebm}
+                {videoAlt}
+            </video>
+            <div class="video-controls controls">
+                <div class="progress-wrap">
+                    <progress class="progress" value="0" max="100">
+                        <span class="progress-bar"></span>
+                    </progress>
+                </div>
+                <div class="buttons">
                     <div class="left-buttons">
-                         <button class="playpause" type="button" data-state="play">Play/Pause</button>
-                         <button class="mute" type="button" data-state="unmute">Mute/Unmute</button>
-                     </div>
-                     <div class="right-buttons">
+                        <button class="playpause" type="button" data-state="play">Play/Pause</button>
+                        <button class="mute" type="button" data-state="unmute">Mute/Unmute</button>
+                    </div>
+                    <div class="right-buttons">
                         <button class="stop" type="button" data-state="stop">Stop</button>
                         <button class="volinc" type="button" data-state="volup">Vol+</button>
                         <button class="voldec" type="button" data-state="voldown">Vol-</button>
                         <button class="fs" type="button" data-state="go-fullscreen">Fullscreen</button>
-                     </div>
                     </div>
                 </div>
-
             </div>
-        </div>';
 
-        $markup .= '</div>';/*.video-wrapper*/
+        </div>
+    </div>
+
+EOD;
+
+    public function getMarkup(array $opts = [])
+    {
+        if (strpos($this->image, '<img') !== false) {
+            preg_match("/\<img.+src\=(?:\"|\')(.+?)(?:\"|\')(?:.+?)\>/", $this->image, $matches);
+            $this->image = $matches[1];
+        }
+        $index        = !empty($opts['index']) ? $opts['index'] : 0;
+        $playerWrapId = 'video-player-wrap-' . $index;
+        $lazy         = (int)$this->lazy === 1;
+        $autoplay     = (int)$this->autoplay === 1;
+        $data         = [
+            '{image}'       => !empty($this->image) ? '<div class="video-image-wrapper"><img src="' . $this->image . '" alt=""></div>' : '',
+            '{label}'       => trad('video.play.label', WWP_THEME_TEXTDOMAIN) . ' ' . $this->title,
+            '{id}'          => $playerWrapId,
+            '{poster}'      => !empty($this->image) ? ' poster="' . $this->image . '" ' : '',
+            '{videoMp4}'    => !empty($this->videoMp4) ? '<source src="' . $this->videoMp4 . '" type="video/mp4">' : '',
+            '{videoOgg}'    => !empty($this->videoOgg) ? '<source src="' . $this->videoOgg . '" type="video/ogg">' : '',
+            '{videoWebm}'   => !empty($this->videoWebm) ? '<source src="' . $this->videoWebm . '" type="video/webm">' : '',
+            '{videoAlt}'    => trad('no_video_support.trad', WWP_THEME_TEXTDOMAIN),
+            '{lazy}'        => $lazy ? 'true' : 'false',
+            '{autoplay}'    => $autoplay ? ' autoplay ' : ' ',
+            '{muted}'       => ' muted="true" ',
+            '{playsinline}' => 'playsinline',
+        ];
+
+        if ($lazy) {
+            $varJs           = 'videoPlayer' . $index;
+            $data['{muted}'] = 'muted="false"';
+            $markup          = '<div class="video-wrapper" data-video-native-component data-video-lazy="true" data-var-name="' . $varJs . '"><script>window.' . $varJs . ' = ' . json_encode([
+                    'template' => static::$template,
+                    'data'     => $data,
+                ]) . '</script></div>';
+        } else {
+            $markup = '<div class="video-wrapper" data-video-native-component>' . str_replace(array_keys($data), array_values($data), self::$template) . '</div>';
+        }
 
         return $markup;
     }
