@@ -2,12 +2,17 @@ import "./Accordion/AccordionComponent"
 import "./Dropdown/DropdownComponent";
 import "./Form/FormComponent";
 import "./Header/HeaderComponent";
-//import "./MapSearch/MapSearchComponent";
 import "./Modal/ModalComponent";
 import "./NoAjaxTransition/NoAjaxTransitionComponent";
 import "./Notification/NotificationComponent";
-import "./Slider/SliderComponent";
-import "./Slider/SliderGallery/SliderGallery";
 import "./Tabs/TabsComponent";
 import "./VideoEmbed/VideoEmbedComponent";
 import "./Video/VideoComponent";
+import "./ToggleContentViaFilter/TcvfComponent";
+
+// SLIDERS
+import "./Slider/SliderDefault/SliderComponent";
+import "./Slider/SliderCarousel/SliderCarouselComponent";
+import "./Slider/SliderGutenbergGallery/SliderGallery";
+
+//import "./MapSearch/MapSearchComponent";
