@@ -10,10 +10,9 @@ namespace WonderWp\Theme\Child\Service;
 
 use WonderWp\Theme\Child\Components\Card\CardComponent;
 use WonderWp\Theme\Child\Components\Dropdown\DropdownComponent;
-use WonderWp\Theme\Child\Components\MapSearch\MapSearchShortcodeHandler;
 use WonderWp\Theme\Child\Components\Modal\ModalComponent;
-use WonderWp\Theme\Child\Components\Slider\SliderComponent;
-use WonderWp\Theme\Child\Components\Slider\SliderItem\SliderItem;
+use WonderWp\Theme\Child\Components\Slider\SliderDefault\SliderComponent;
+use WonderWp\Theme\Child\Components\Slider\SliderDefault\SliderItem\SliderItem;
 use WonderWp\Theme\Child\Components\Tabs\TabsComponent;
 use WonderWp\Theme\Child\Components\Tabs\TabItem\TabItem;
 use WonderWp\Theme\Child\Components\Timeline\TimelineComponent;
@@ -23,9 +22,9 @@ use WonderWp\Theme\Core\Service\ThemeShortcodeService;
 class ChildThemeShortcodeService extends ThemeShortcodeService
 {
 
-    public function registerShortcodes()
+    public function register()
     {
-        parent::registerShortcodes();
+        parent::register();
 
         add_shortcode('slider', [$this, 'slider']);
         add_shortcode('slider-item', [$this, 'slideritem']);
@@ -37,6 +36,7 @@ class ChildThemeShortcodeService extends ThemeShortcodeService
         add_shortcode('tabs', [$this, 'tabs']);
         add_shortcode('tab-item', [$this, 'tabitem']);
         //add_shortcode(MapSearchShortcodeHandler::$shortCode, [MapSearchShortcodeHandler::class, 'handle']);
+        add_shortcode('social-share', [$this, 'socialShare']);
 
         return $this;
     }
@@ -134,20 +134,17 @@ class ChildThemeShortcodeService extends ThemeShortcodeService
 
     public function tabs($attr, $content)
     {
-        $tab = new TabsComponent();
-
-        $nbTabs   = $attr['nbtabs'];
-        $tabItems = [];
+        $tabComponent = new TabsComponent();
 
         $shortcodes = $this->extractShortcodes($content, 'tab-item');
+        $nbTabs     = count($shortcodes);
 
         for ($i = 1; $i <= $nbTabs; $i++) {
-            $tabItems[$i]['markup'] = do_shortcode($shortcodes[$i - 1]);
-            $tabItems[$i]['title']  = $attr['title_' . $i];
+            $res = json_decode(do_shortcode($shortcodes[$i - 1]));
+            $tabComponent->addTab($res->title, $res->content, $res->id, $res->class);
         }
-        $tab->tabItems = $tabItems;
 
-        return $tab->getMarkup();
+        return $tabComponent->getMarkup();
     }
 
     public function tabitem($attr, $content)
@@ -159,7 +156,7 @@ class ChildThemeShortcodeService extends ThemeShortcodeService
             $tabitem->content = $content;
         }
 
-        return $tabitem->getMarkup();
+        return json_encode($tabitem, true);
     }
 
     public static function extractShortcode($content, $element)
