@@ -1,11 +1,14 @@
 <!-- components/base/fonts.php -->
 
-<p class="font font1">
+<p>Privilégier si possible des Google Fonts. L'import doit être fait dans le fichier variables.scss</p>
+<p>Si des fonts natives sont utilisées, elles devront être chargées dans le répertoire : /web/app/themes/wwp_child_theme/assets/fonts.</p>
+
+<p class="font font-base">
     <span>Base font example : $font-family-base</span>
 </p>
-<p class="font font2">
+<p class="font font-headings">
     <span>Headings font example : $font-family-headings</span>
 </p>
-<p class="font font3">
+<p class="font font-secondary">
     <span>Secondary font example : $font-family-secondary</span>
 </p>
