@@ -1,9 +1,14 @@
 <!-- components/base/images.php -->
 
-<p class="subtitle">Image à la une détail actualité : 1730x530px</p>
-<img width="1730" height="530" src="https://via.placeholder.com/1730x530" alt="">
+<p class="subtitle">Image à la une : 1920x1080</p>
+<img width="1920" height="1080" src="https://via.placeholder.com/1920x1080" alt="">
 <hr>
 
-<p class="subtitle">Image thumbnail liste actualités : 298x450px</p>
-<img width="298" height="450" src="https://via.placeholder.com/298x450" alt="">
+<p class="subtitle">Image thumbnail sur 1 colonne : 293x220px</p>
+<img width="293" height="220" src="https://via.placeholder.com/293x220" alt="">
 <hr>
+
+<p class="subtitle">Image medium sur 2 colonnes : 460x345px</p>
+<img width="460" height="345" src="https://via.placeholder.com/460x345" alt="">
+<hr>
+
