@@ -5,7 +5,7 @@ $video = new \WonderWp\Theme\Child\Components\Video\VideoComponent();
 
 $video
     ->setTitle('Titre de la video')
-    ->setImage('http://via.placeholder.com/920x700')
+    ->setImage('https://placeimg.com/340/250/nature')
     ->setVideoMp4('/app/themes/wwp_child_theme/assets/raw/videos/clown.mp4')
     ->setVideoOgg('/app/themes/wwp_child_theme/assets/raw/videos/clown.ogv')
     ->setVideoWebm('/app/themes/wwp_child_theme/assets/raw/videos/clown.webm')
